requirements.txt
# Comic Generation SaaS - Database Layer Requirements

# Core dependencies
beanie>=1.23.0          # MongoDB ODM built on Pydantic
motor>=3.3.0            # Async MongoDB driver
pydantic>=2.0.0         # Data validation (v2 required)
pymongo>=4.6.0          # MongoDB driver (required by Beanie)

# Authentication & Security
python-jose[cryptography]>=3.3.0    # JWT handling
passlib[bcrypt]>=1.7.4              # Password hashing with bcrypt
email-validator>=2.0.0              # Email validation for Pydantic

# FastAPI & Web
fastapi>=0.104.0                    # Web framework
uvicorn[standard]>=0.24.0           # ASGI server
python-multipart>=0.0.6             # Form data parsing

# LLM & Networking
httpx>=0.25.0                       # Async HTTP client
tenacity>=8.2.0                     # Retry logic

# Async Tasks
celery>=5.3.0                       # Task queue
redis>=5.0.0                        # Broker & Backend
asgiref>=3.7.0                      # Async-Sync utilities

